requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
pydantic-settings
redis
rq
transformers
torch
numpy
qdrant-client
langchain
langchain-openai
python-jose[cryptography]
gunicorn
langchain_community
pyjwt